TDD Search step: refactoring
